fix: DB prices in integer

+++ database/migrations/2024_03_04_094032_create_discounts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->string('code')->unique();
-            $table->float('amount');
+            $table->integer('amount');
             $table->timestamp('ended_at');
             $table->timestamps();
         });
